Refactor: Rename "App" to "Card"
We want to add some more pages, so the main component can't be thought
of as the whole app anymore.
