+++ src/components/Card.test.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { render, fireEvent, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import App from "./App";
+import Card from "./Card";
 import stripIndent from "strip-indent";
 
 describe("'New card' button", () => {
   test("changes the words on the card", () => {
-    render(<App />);
+    render(<Card />);
     const initialCells = screen.queryAllByRole("gridcell");
     const firstWord = initialCells[0].textContent;
     const secondWord = initialCells[1].textContent;
@@ -19,7 +19,7 @@ describe("'New card' button", () => {
   });
 
   test("clears any stamped cells", () => {
-    render(<App />);
+    render(<Card />);
     const cells = screen.queryAllByRole("gridcell");
 
     fireEvent.click(cells[0]);
@@ -34,7 +34,7 @@ describe("'New card' button", () => {
 
 describe("'Clear' button", () => {
   test("clears any stamped cells", () => {
-    render(<App />);
+    render(<Card />);
     const cells = screen.queryAllByRole("gridcell");
     const firstCell = cells[0];
     const lastCell = cells[24];
@@ -57,7 +57,7 @@ describe("'Share' button", () => {
     const user = userEvent.setup();
     render(
       <>
-        <App />
+        <Card />
         <textarea rows={5} />
       </>
     );
